refactor: extract DiagnosticsLogger.swift from BugNarratorDiagnostics.swift
Extracts DiagnosticsLogger struct (75 lines) — the logger implementation used
by all diagnostic categories — from BugNarratorDiagnostics.swift into its own
file.

Byte-identical move; SHA-256 to be recomputed against origin/main lines 203-277
(post-#673 baseline).

Imports: OSLog (for Logger + privacy) + Foundation.

Reduces BugNarratorDiagnostics.swift by 75 lines.

Tests: 667 (unchanged).
